Reorganise view layers recoil state
