+++ containers/web-server/frontend/src/state/layers/data-layers/buildings.ts
@@ -1,0 +1,16 @@
+import { selector } from 'recoil';
+
+import { ViewLayer } from '@/lib/data-map/view-layers';
+import { truthyKeys } from '@/lib/helpers';
+
+import { buildingsViewLayer } from '@/config/buildings/buildings-view-layer';
+import { buildingSelectionState } from '@/state/buildings';
+import { sectionVisibilityState } from '@/state/sections';
+
+export const buildingLayersState = selector<ViewLayer[]>({
+  key: 'buildingLayersState',
+  get: ({ get }) =>
+    get(sectionVisibilityState('buildings'))
+      ? truthyKeys(get(buildingSelectionState)).map((buildingType) => buildingsViewLayer(buildingType))
+      : [],
+});
