+++ containers/web-server/frontend/src/state/layers/ui-layers/feature-bbox.ts
@@ -1,0 +1,23 @@
+import bboxPolygon from '@turf/bbox-polygon';
+import { selector } from 'recoil';
+
+import { extendBbox } from '@/lib/bounding-box';
+import { ViewLayer, viewOnlyLayer } from '@/lib/data-map/view-layers';
+import { boundingBoxLayer } from '@/lib/deck/layers/bounding-box-layer';
+
+import { hoveredAdaptationFeatureState } from '@/details/adaptations/FeatureAdaptationsTable';
+
+export const featureBoundingBoxLayerState = selector<ViewLayer>({
+  key: 'featureBoundingBoxLayerState',
+  get: ({ get }) => {
+    const hoveredAdaptationFeature = get(hoveredAdaptationFeatureState);
+
+    if (!hoveredAdaptationFeature) return null;
+
+    const geom = bboxPolygon(extendBbox(hoveredAdaptationFeature.bbox, 5));
+
+    return viewOnlyLayer(`feature-bounding-box-${hoveredAdaptationFeature.id}`, ({ deckProps }) =>
+      boundingBoxLayer({ bboxGeom: geom }, deckProps),
+    );
+  },
+});
